Issue #27748 split experimental function out of Primes repository
