+++ dev/io.openliberty.data.internal_fat/bnd.bnd
@@ -1,5 +1,5 @@
 #*******************************************************************************
-# Copyright (c) 2022,2023 IBM Corporation and others.
+# Copyright (c) 2022,2024 IBM Corporation and others.
 # All rights reserved. This program and the accompanying materials
 # are made available under the terms of the Eclipse Public License 2.0
 # which accompanies this distribution, and is available at
@@ -38,20 +38,13 @@ tested.features: \
 -sub: *.bnd
 
 -buildpath: \
-  com.ibm.tx.jta.jakarta;version=latest,\
-  com.ibm.websphere.appserver.spi.kernel.service;version=latest,\
   com.ibm.websphere.appserver.spi.logging,\
   com.ibm.websphere.org.osgi.core,\
   com.ibm.websphere.org.osgi.service.component,\
   com.ibm.wsspi.org.osgi.service.component.annotations,\
-  com.ibm.ws.cdi.interfaces.jakarta;version=latest, \
   com.ibm.ws.componenttest.2.0;version=latest,\
-  com.ibm.ws.kernel.service;version=latest,\
   com.ibm.ws.org.osgi.annotation.versioning;version=latest,\
-  com.ibm.ws.persistence.jakarta;version=latest,\
-  com.ibm.ws.tx.embeddable.jakarta;version=latest,\
   fattest.simplicity;version=latest,\
-  io.openliberty.data,\
   io.openliberty.jakarta.annotation.2.1,\
   io.openliberty.jakarta.cdi.4.0,\
   io.openliberty.jakarta.concurrency.3.0,\
